refactor storage
